added delete project

+++ src/app/pages/project/project.component.ts
@@ -153,7 +153,16 @@ export class ProjectComponent implements OnInit {
     if (!this.ngoStorage.ngo || !this.project) {
       return;
     }
+    if (!confirm('Are you sure you want to delete this project?')){
+      return;
+    }
     this.apiService.project.deleteProject(this.ngoStorage.ngo.id, this.project.id).then(()=> {
+      const removed = this.ngoStorage.ngo?.projects.projects.filter(p => p.id !== this.project?.id);
+      const projects = this.ngoStorage.ngo?.projects;
+      if(projects && removed){
+        projects.projects = removed;
+      }
+      this.router.navigate(['/overview']).then();
     })
   }
